+++ .eslintrc.js
@@ -21,7 +21,7 @@ module.exports = {
     project: true,
   },
   rules: {
-    '@typescript-eslint/non-null-assertion': 'off',
+    '@typescript-eslint/no-non-null-assertion': 'off',
     '@typescript-eslint/no-explicit-any': 'off',
   },
 }