+++ src/test/suite/index.ts
@@ -5,11 +5,12 @@ import { glob } from 'glob'
 export async function run() {
   // Create the mocha test
   const mocha = new Mocha({
-    ui: 'tdd',
+    ui: 'bdd',
     color: true,
   })
   const testsRoot = path.resolve(__dirname, '..')
   const files = await glob('**/*.test.js', { cwd: testsRoot })
+  console.log('files:', files)
 
   // Add files to the test suite
   files.forEach((f) => mocha.addFile(path.resolve(testsRoot, f)))
@@ -25,3 +26,5 @@ export async function run() {
     throw err
   }
 }
+
+void run()